requirements.txt
asgiref==3.3.0
certifi==2020.6.20
cloudinary==1.22.0
dj-database-url==0.5.0
Django==3.1.3
djangorestframework==3.12.1
psycopg2-binary==2.8.6
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
